pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aigym"
version = "0.1.0"
description = "AIGYM — handshake & interoperability protocol for AI↔AI"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{name = "AIGYM maintainers"}]
dependencies = ["pyyaml>=6.0.1","jsonschema>=4.22.0"]

[project.scripts]
aigym = "aigym.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
[tool.setuptools.package-data]
aigym = ["schema/*.json"]
